fix: refactor App component to separate file
